#360 improved matching api
