+++ profiler/src/main/java/com/navercorp/pinpoint/profiler/modifier/DefaultModifierRegistry.java
@@ -17,22 +17,16 @@
 package com.navercorp.pinpoint.profiler.modifier;
 
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 import com.navercorp.pinpoint.bootstrap.Agent;
 import com.navercorp.pinpoint.bootstrap.config.ProfilerConfig;
 import com.navercorp.pinpoint.bootstrap.instrument.ByteCodeInstrumentor;
 import com.navercorp.pinpoint.bootstrap.instrument.matcher.ClassNameMatcher;
+import com.navercorp.pinpoint.bootstrap.instrument.matcher.MultiClassNameMatcher;
 import com.navercorp.pinpoint.bootstrap.instrument.matcher.Matcher;
-import com.navercorp.pinpoint.profiler.modifier.arcus.ArcusClientModifier;
-import com.navercorp.pinpoint.profiler.modifier.arcus.BaseOperationModifier;
-import com.navercorp.pinpoint.profiler.modifier.arcus.CacheManagerModifier;
-import com.navercorp.pinpoint.profiler.modifier.arcus.CollectionFutureModifier;
-import com.navercorp.pinpoint.profiler.modifier.arcus.FrontCacheGetFutureModifier;
-import com.navercorp.pinpoint.profiler.modifier.arcus.GetFutureModifier;
-import com.navercorp.pinpoint.profiler.modifier.arcus.ImmediateFutureModifier;
-import com.navercorp.pinpoint.profiler.modifier.arcus.MemcachedClientModifier;
-import com.navercorp.pinpoint.profiler.modifier.arcus.OperationFutureModifier;
+import com.navercorp.pinpoint.profiler.modifier.arcus.*;
 import com.navercorp.pinpoint.profiler.modifier.connector.asynchttpclient.AsyncHttpClientModifier;
 import com.navercorp.pinpoint.profiler.modifier.connector.httpclient3.DefaultHttpMethodRetryHandlerModifier;
 import com.navercorp.pinpoint.profiler.modifier.connector.httpclient3.HttpClientModifier;
@@ -57,9 +51,7 @@ import com.navercorp.pinpoint.profiler.modifier.db.mysql.MySQLPreparedStatementM
 import com.navercorp.pinpoint.profiler.modifier.db.mysql.MySQLStatementModifier;
 import com.navercorp.pinpoint.profiler.modifier.db.oracle.OracleDriverModifier;
 import com.navercorp.pinpoint.profiler.modifier.db.oracle.OraclePreparedStatementModifier;
-import com.navercorp.pinpoint.profiler.modifier.db.oracle.OraclePreparedStatementWrapperModifier;
 import com.navercorp.pinpoint.profiler.modifier.db.oracle.OracleStatementModifier;
-import com.navercorp.pinpoint.profiler.modifier.db.oracle.OracleStatementWrapperModifier;
 import com.navercorp.pinpoint.profiler.modifier.db.oracle.PhysicalConnectionModifier;
 import com.navercorp.pinpoint.profiler.modifier.log.log4j.LoggingEventOfLog4jModifier;
 import com.navercorp.pinpoint.profiler.modifier.log.logback.LoggingEventOfLogbackModifier;
@@ -103,12 +95,22 @@ public class DefaultModifierRegistry implements ModifierRegistry {
 
     public void addModifier(AbstractModifier modifier) {
         final Matcher matcher = modifier.getMatcher();
+        // TODO extract matcher process
         if (matcher instanceof ClassNameMatcher) {
             final ClassNameMatcher classNameMatcher = (ClassNameMatcher)matcher;
             String className = classNameMatcher.getClassName();
             AbstractModifier old = registry.put(className, modifier);
             if (old != null) {
                 throw new IllegalStateException("Modifier already exist new:" + modifier.getClass() + " old:" + old.getMatcher());
+            }
+        } else if (matcher instanceof MultiClassNameMatcher) {
+            final MultiClassNameMatcher classNameMatcher = (MultiClassNameMatcher)matcher;
+            List<String> classNameList = classNameMatcher.getClassNames();
+            for (String className : classNameList) {
+                AbstractModifier old = registry.put(className, modifier);
+                if (old != null) {
+                    throw new IllegalStateException("Modifier already exist new:" + modifier.getClass() + " old:" + old.getMatcher());
+                }
             }
         } else {
             throw new IllegalArgumentException("unsupported matcher :" + matcher);
@@ -173,20 +175,14 @@ public class DefaultModifierRegistry implements ModifierRegistry {
                 ArcusClientModifier arcusClientModifier = new ArcusClientModifier(byteCodeInstrumentor, agent);
                 addModifier(arcusClientModifier);
                 // Future of Arcus
-                CollectionFutureModifier collectionFutureModifier = new CollectionFutureModifier(byteCodeInstrumentor, agent);
-                addModifier(collectionFutureModifier);
+//                CollectionFutureModifier collectionFutureModifier = new CollectionFutureModifier(byteCodeInstrumentor, agent);
+//                addModifier(collectionFutureModifier);
             }
 
             // future modifier start ---------------------------------------------------
-
-            GetFutureModifier getFutureModifier = new GetFutureModifier(byteCodeInstrumentor, agent);
+            // unsupport CollectionFutureModifier
+            FutureModifier getFutureModifier = new FutureModifier(byteCodeInstrumentor, agent);
             addModifier(getFutureModifier);
-
-            ImmediateFutureModifier immediateFutureModifier = new ImmediateFutureModifier(byteCodeInstrumentor, agent);
-            addModifier(immediateFutureModifier);
-
-            OperationFutureModifier operationFutureModifier = new OperationFutureModifier(byteCodeInstrumentor, agent);
-            addModifier(operationFutureModifier);
 
 //            Not working properly. commented out for now.
             FrontCacheGetFutureModifier frontCacheGetFutureModifier = new FrontCacheGetFutureModifier(byteCodeInstrumentor, agent);
@@ -309,13 +305,9 @@ public class DefaultModifierRegistry implements ModifierRegistry {
         AbstractModifier oracleConnectionModifier = new PhysicalConnectionModifier(byteCodeInstrumentor, agent);
         addModifier(oracleConnectionModifier);
 
-        AbstractModifier oraclePreparedStatementWrapperModifier = new OraclePreparedStatementWrapperModifier(byteCodeInstrumentor, agent);
-        addModifier(oraclePreparedStatementWrapperModifier);
         AbstractModifier oraclePreparedStatementModifier = new OraclePreparedStatementModifier(byteCodeInstrumentor, agent);
         addModifier(oraclePreparedStatementModifier);
 
-        AbstractModifier oracleStatementWrapperModifier = new OracleStatementWrapperModifier(byteCodeInstrumentor, agent);
-        addModifier(oracleStatementWrapperModifier);
         AbstractModifier oracleStatementModifier = new OracleStatementModifier(byteCodeInstrumentor, agent);
         addModifier(oracleStatementModifier);
         //
