--- profiler/src/main/java/com/navercorp/pinpoint/profiler/modifier/db/oracle/OraclePreparedStatementWrapperModifier.java
@@ -1,49 +1,0 @@
-/*
- * Copyright 2014 NAVER Corp.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.navercorp.pinpoint.profiler.modifier.db.oracle;
-
-import java.security.ProtectionDomain;
-
-import com.navercorp.pinpoint.bootstrap.Agent;
-import com.navercorp.pinpoint.bootstrap.instrument.ByteCodeInstrumentor;
-import com.navercorp.pinpoint.bootstrap.instrument.matcher.Matcher;
-import com.navercorp.pinpoint.bootstrap.instrument.matcher.Matchers;
-import com.navercorp.pinpoint.profiler.modifier.AbstractModifier;
-import com.navercorp.pinpoint.profiler.modifier.ModifierDelegate;
-
-/**
- * @author emeroad
- */
-public class OraclePreparedStatementWrapperModifier extends AbstractModifier {
-
-    private final ModifierDelegate delegate;
-
-    public OraclePreparedStatementWrapperModifier(ByteCodeInstrumentor byteCodeInstrumentor, Agent agent) {
-        super(byteCodeInstrumentor, agent);
-        this.delegate = new OraclePreparedStatementModifierDelegate(byteCodeInstrumentor);
-    }
-
-    public Matcher getMatcher() {
-        return Matchers.newClassNameMatcher(OracleClassConstants.ORACLE_PREPARED_STATEMENT_WRAPPER);
-    }
-
-    @Override
-    public byte[] modify(ClassLoader classLoader, String javassistClassName, ProtectionDomain protectedDomain, byte[] classFileBuffer) {
-        return this.delegate.modify(classLoader, javassistClassName, protectedDomain, classFileBuffer);
-    }
-
-}
