+++ profiler/src/main/java/com/navercorp/pinpoint/profiler/modifier/db/oracle/OracleStatementModifier.java
@@ -20,10 +20,17 @@ import java.security.ProtectionDomain;
 
 import com.navercorp.pinpoint.bootstrap.Agent;
 import com.navercorp.pinpoint.bootstrap.instrument.ByteCodeInstrumentor;
+import com.navercorp.pinpoint.bootstrap.instrument.InstrumentClass;
+import com.navercorp.pinpoint.bootstrap.instrument.InstrumentException;
 import com.navercorp.pinpoint.bootstrap.instrument.matcher.Matcher;
 import com.navercorp.pinpoint.bootstrap.instrument.matcher.Matchers;
+import com.navercorp.pinpoint.bootstrap.interceptor.Interceptor;
+import com.navercorp.pinpoint.bootstrap.interceptor.tracevalue.DatabaseInfoTraceValue;
 import com.navercorp.pinpoint.profiler.modifier.AbstractModifier;
-import com.navercorp.pinpoint.profiler.modifier.ModifierDelegate;
+import com.navercorp.pinpoint.profiler.modifier.db.interceptor.StatementExecuteQueryInterceptor;
+import com.navercorp.pinpoint.profiler.modifier.db.interceptor.StatementExecuteUpdateInterceptor;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
  * For ojdbc library without OracleStatementWrapper.
@@ -33,25 +40,58 @@ import com.navercorp.pinpoint.profiler.modifier.ModifierDelegate;
  */
 public class OracleStatementModifier extends AbstractModifier {
 
-    private final ModifierDelegate delegate;
+    private final Logger logger = LoggerFactory.getLogger(this.getClass());
     
     public OracleStatementModifier(ByteCodeInstrumentor byteCodeInstrumentor, Agent agent) {
         super(byteCodeInstrumentor, agent);
-        this.delegate = new OracleStatementModifierDelegate(byteCodeInstrumentor);
     }
 
     @Override
     public Matcher getMatcher() {
-        return Matchers.newClassNameMatcher(OracleClassConstants.ORACLE_STATEMENT);
+        return Matchers.newMultiClassNameMatcher(OracleClassConstants.ORACLE_STATEMENT, OracleClassConstants.ORACLE_STATEMENT_WRAPPER);
     }
 
     @Override
     public byte[] modify(ClassLoader classLoader, String javassistClassName, ProtectionDomain protectedDomain, byte[] classFileBuffer) {
      // Do not modify if wrapper exists
-        if (byteCodeInstrumentor.findClass(classLoader, OracleClassConstants.ORACLE_STATEMENT_WRAPPER)) {
+        if(OracleClassConstants.ORACLE_STATEMENT.equals(javassistClassName)) {
+            if (byteCodeInstrumentor.findClass(classLoader, OracleClassConstants.ORACLE_STATEMENT_WRAPPER)) {
+                return null;
+            }
+        }
+        return this.modifyStatement(classLoader, javassistClassName, protectedDomain, classFileBuffer);
+    }
+
+    public byte[] modifyStatement(ClassLoader classLoader, String className, ProtectionDomain protectedDomain, byte[] classFileBuffer) {
+        if (logger.isInfoEnabled()) {
+            logger.info("Modifying. {}", className);
+        }
+        try {
+            InstrumentClass statementClass = byteCodeInstrumentor.getClass(classLoader, className, classFileBuffer);
+            Interceptor executeQuery = new StatementExecuteQueryInterceptor();
+            statementClass.addGroupInterceptor("executeQuery", new String[]{"java.lang.String"}, executeQuery, OracleScope.SCOPE_NAME);
+
+            // FIXME
+            Interceptor executeUpdateInterceptor1 = new StatementExecuteUpdateInterceptor();
+            statementClass.addGroupInterceptor("executeUpdate", new String[]{"java.lang.String"}, executeUpdateInterceptor1, OracleScope.SCOPE_NAME);
+
+
+            Interceptor executeUpdateInterceptor2 = new StatementExecuteUpdateInterceptor();
+            statementClass.addGroupInterceptor("executeUpdate", new String[]{"java.lang.String", "int"}, executeUpdateInterceptor2, OracleScope.SCOPE_NAME);
+
+            Interceptor executeInterceptor1 = new StatementExecuteUpdateInterceptor();
+            statementClass.addGroupInterceptor("execute", new String[]{"java.lang.String"}, executeInterceptor1, OracleScope.SCOPE_NAME);
+
+            Interceptor executeInterceptor2 = new StatementExecuteUpdateInterceptor();
+            statementClass.addGroupInterceptor("execute", new String[]{"java.lang.String", "int"}, executeInterceptor2, OracleScope.SCOPE_NAME);
+
+            statementClass.addTraceValue(DatabaseInfoTraceValue.class);
+            return statementClass.toBytecode();
+        } catch (InstrumentException e) {
+            if (logger.isWarnEnabled()) {
+                logger.warn("{} modify fail. Cause:{}", this.getClass().getSimpleName(), e.getMessage(), e);
+            }
             return null;
         }
-        return this.delegate.modify(classLoader, javassistClassName, protectedDomain, classFileBuffer);
     }
-    
 }
